cat-file: get rid of expand_data
Clean up cat-file after moving all formatting logic
to ref-filter.
We do not need to use struct expand_data anymore.
